stocks/requirements.txt
yahoo_finance
Google-Search-API
wikipedia
python-dateutil
Flask
flask-debugtoolbar
flask-sqlalchemy